feat: change int to int64

+++ Sources/SolanaSwift/Programs/TokenPrograms/Token2022Program/ParsedTransferFeeConfig.swift
@@ -4,12 +4,12 @@ public struct ParsedTransferFeeConfig: Codable {
     public static let name: String = "transferFeeConfig"
 
     public struct TransferFee: Codable {
-        let epoch: Int
+        let epoch: Int64
         let maximumFee: Double
-        let transferFeeBasisPoints: Int
+        let transferFeeBasisPoints: Int64
     }
 
     public let newerTransferFee, olderTransferFee: TransferFee
     public let transferFeeConfigAuthority, withdrawWithheldAuthority: String
-    public let withheldAmount: Int
+    public let withheldAmount: Int64
 }
